move completions to a new mod in ruff crate show all option sets, even without a doc do not show completion values in help
