+++ crates/ruff/Cargo.toml
@@ -25,7 +25,7 @@ ruff_python_formatter = { workspace = true }
 ruff_server = { workspace = true }
 ruff_source_file = { workspace = true }
 ruff_text_size = { workspace = true }
-ruff_workspace = { workspace = true, features = ["clap"] }
+ruff_workspace = { workspace = true }
 
 anyhow = { workspace = true }
 argfile = { workspace = true }
