+++ crates/ruff/src/commands/completions/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod config;
