+++ crates/ruff_workspace/Cargo.toml
@@ -21,7 +21,7 @@ ruff_macros = { workspace = true }
 ruff_python_ast = { workspace = true }
 ruff_python_formatter = { workspace = true, features = ["serde"] }
 ruff_python_semantic = { workspace = true, features = ["serde"] }
-ruff_python_stdlib = { workspace = true }
+ruff_python_stdlib = {workspace = true}
 ruff_source_file = { workspace = true }
 
 anyhow = { workspace = true }
@@ -43,7 +43,6 @@ serde = { workspace = true }
 shellexpand = { workspace = true }
 strum = { workspace = true }
 toml = { workspace = true }
-clap = { workspace = true, features = ["derive", "string"], optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 etcetera = { workspace = true }
@@ -59,12 +58,7 @@ ignored = ["colored"]
 
 [features]
 default = []
-schemars = [
-    "dep:schemars",
-    "ruff_formatter/schemars",
-    "ruff_python_formatter/schemars",
-    "ruff_python_semantic/schemars",
-]
+schemars = ["dep:schemars", "ruff_formatter/schemars", "ruff_python_formatter/schemars", "ruff_python_semantic/schemars"]
 
 [lints]
 workspace = true
