+++ crates/ruff_workspace/src/lib.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "clap")]
-pub mod cli;
 pub mod configuration;
 pub mod options;
 pub mod pyproject;
